build(backend): add pgadmin to docker-compose
The pgadmin container is added to the docker-compose file to allow
for easy management of the PostgreSQL database. The pgadmin service
is configured to use the same network as the backend and PostgreSQL
services, and it exposes port 5050 for access.
Plus the example.env file is updated to include the pgadmin configuration
and the default email and password for the admin user